Split remaining 1000+ line files into focused modules
